pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["v.alehin <vladislav.alehin@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = "^0.29.0"
fastapi = "^0.111.0"
pydantic = "^2.7.1"
jinja2 = "^3.1.4"
sqlalchemy = "^2.0.30"
alembic = "^1.13.1"
tenacity = "^8.3.0"
greenlet = "^3.0.3"
websocket = "^0.2.1"
fastapi-csrf-protect = "^0.3.3"
passlib = "^1.7.4"
pyjwt = "^2.8.0"
gunicorn = "^23.0.0"
grpcio = "^1.70.0"
grpcio-tools = "^1.70.0"
tinode-grpc = "^0.22.3"
pymongo = "^4.11"
ruff = "^0.9.5"
black = "^25.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
